Remove padding from around the outside of SVGs
SVG symbols had a lot of padding within the path of the SVG itself. Now it's a minimal amount which just allows the symbols to be centered.

Since every symbol was being changed, I decided it was also a good opportunity to alphabetically sort the symbols.

Additionally, lock and unlock symbols have been added.