Add styling for new SVG icons to xui
This commit adds styles for the new SVGs, and changes `xui-button-icon` to work with the new icons.

`xui-icon-svg` is the styling class applied to icons within the icon blob, `xui-icon` or the `-lg` and `-xlg` variants
are added to icons when inserted into a document.

Three sizes are provided, standard, large and x-large. No small variant has been provided as of yet due to the rendering issues the current icons have
when decreased below their original size. Additionally, the `xui-button-icon-sm` class has been removed.