Fix XUI styles in Positioned (Portaled) components
React Portal used for positioning is rendered at the bottom of the DOM.
If this occurs on a page that does not have `xui-body` applied it
will correctly style children.

This PR adds `xui-container` (our wrapper class for using XUI on hybrid
pages where applying `xui-body` is not desired) to a portal container
element. Ensuring base XUI styles are available.

ref: none