Update umd file names
This CSS file bundles component specific classes so it is different from XUI on the CDN. I've added
umd into the CSS file name to reflect it goes with the umd bundle.

Brought in the release version No. from package.json and used in file name in preperation for
publishing to AF or S3 or attaching to the GH release.

ref: none

cc @dev-johnsanders @dev-williamO @UXE/uxe-team