Modify Positioning to use translateX over left css property (#1983)
* Modify Positioning to use translateX over left css property

Using left was causing elements to occasionally be rendered too small if the trigger was to the right side of the screen. I think this is because their width was calculated as what could fit between the existing left attribute and the right side of the screen. Since translateX doesn't actually affect the layout, it doesn't affect the width calculation (it's also more performant)

* Modify to use math.floor to avoid attempt at subpixel rendering

* Change default transform to null