Fix Autocompleter open on focus bug for IE11 (#2090)
* Update README

* Update README

* Fix Autocompleter open on focus bug for IE11

IE11 would automatically re-trigger a focus event and re-open the dropdown on close/blur
making it impossible to close.

I've not found the root cause, but suspect it is around a re-render being triggered withthe
prop causing it to pop right open.

This fix adds a check to the  handler added when the  is passed in as a prop.
It checks to see if the component is already focused before triggering the call to .

This fixes the bug with closing for IE11.

The behaviour on IE11 varies slightly. When a click away blur happens, IE retains focus in the input box
where Chrome, FF, Edge clear the focus. This causes IE to not trigger an open on focus until the input has
been blurred first. Other browsers will trigger the open on the next click of the input (because it has lost focus)

This behaviour seems acceptable.

Tested: FF, Chrome Mac. Edge, IE11 Windows.

fixes: #2081

ref: @UXE/uxe-team @jishaal