Fix styling of current styleguide when deprecated classes are removed. (#2175)
This PR temporarily moves some styles from XUI into kss under the  namespace.
A bunch of ~/dev/uxe/xui classes are removed from kss.

This isn't exact. But all of this is being re-factoered currently. And this PR will
enable the docs to work in the immediate term.

ref: none.

cc @UXE/uxe-team