Add consistent validation layout with React examples (#2466)
* Add consistent validation layout with React examples

Validation layout should be consistent regardless of whether a message is currently being shown or not. This uses position: absolute to consistenlty position the message, and applies xui-field-layout via a prop on XUIStatelesInput. Since this problem is only apparnet in interactive examples, I've updated the validation example in the react docs to show this behaviour

* Change to use negative bottom margin