Add flex-grow to div wrapping scrollable content and footer in dropdown (#2744)
The div for scrollable content has flex-grow:1 which enables it to grow to the height of it's container if it has a set height. But the wrapping div did not. This meant if DropDown was given a set height greater than it's contents, the footer would sit right under the height of the scrollable contents, leaving space at the bottom of the dropdown.