abstract out XUI var usage to enable landing the color swatch changes (#3635)
This commit focuses on usages of greys within the design system. The
goal is to abstract them out so when the colors all change there is only
20 or so variables to update all in one place.

This PR also separates the variables into their own file. This is mostly
for example.scss which is pulled is built for storybook and needs to
have all it's deps explicitly available. DOnt worry.

ref: #3057