[XUI-293] Add size option for SelectBox
[XUI-293] Add size option for SelectBox

The size option is passed through to the XUIButton that is used for the
SelectBox trigger. The functionality that `xui-select--button` has been split
out to make sure that the button size is not overridden by `xui-select--button`.

Making the CSS for the SelectBox more modular also allowed me to decouple the
`isTextTruncated` prop, which means it can be turned off even if a
`buttonVariant` isn't supplied.