[XUI-261] NPM audit fixes
This fixes *most* of the vulnerabilities revealed by `npm audit`. One of
those fixes necessitated a Jest update, which necessitated tweaking the
Jest config to silence a deprecation warning.

Stylelint was picking up a ton of whitespace errors in temp files
generated by KSS, so I took the liberty of ignoring those and fixing the
one actual gremlin in `_overview.scss`.

Storybook leaves two low and two moderate vulnerabilities behind, but
the major upgrade to fix those is out of scope for this ticket.