[XUI-446] Fix stepper tab truncation in IE11
This issue is similar to the one in #4293, only rotated 90 degrees.
Adding a `max-width` to the offending element’s child seems enough to
force IE11 to calculate the width of the element correctly.

The `max-width` in question is a magic number internal to the component.
I figured needing to add this in one more place was enough to
necessitate creating a Sass variable for it.