[XUI-395] Improve XUIStepper QA hooks
Teams implementing `XUIStepper` were having to reach in to the component
to grab the button on each tab. This exposes the buttons in the active
stepper layout while leaving the buttons in the dummy layouts hidden.

Per user request, I also added a `data-` attribute to indicate if the
button belongs to an active tab.