[patch] XUI-756 Fix KSS build regression
A recent upgrade to the KSS version regressed the Docs. New KSS supports different builders and the code to resolve which builder no longer resolved our custom one.

It seems it was simply an issue with resolving the path. This PR removes the current directory from the config path which seems to work in the KSS path resolver function.

The other KSS paths in the config have been untested.

Fixes: XUI-756