[XUI-808] Make tabIndex prop types friendlier
Our `tabIndex` prop types were a mishmosh of string, number, or
string-or-number. I changed *most* of them to number in #4391, but we
decided that allowing a string or a number was user-friendlier. I went
ahead and changed all the affected prop types here, not just the ones I
had previously touched.