[breaking-changes] [XUI-842] Page and isolation header CSS tweaks (#4458)
* [XUI-842] Page and isolation header CSS tweaks

This boosts the secondary header font size and tweaks the spacing
between elements in the title wrapper. This required moving around some
furniture – to avoid strange effects in the isolation header, I moved
the spacing from the left side of the affected elements to the right
side and adjusted the left-content margin accordingly.

I also rearranged some of the mixin to make the children easier to find
and deleted some zombie isolation-header variables.

* [XUI-842] Fix CSS override for Styleguidist

Changing the isolation header class broke an override in Styleguidist.
Everything else remained unaffected, but this fixes it everywhere.