[breaking-changes] [CHORE] Add utility scripts and CSS watcher (#4451)
* [CHORE] Add utility scripts to package.json

This adds additional tasks to build and run Storybook and Styleguidist,
as well as a couple test aliases and a `review` command for pre-PR
proofing. It doesn’t look like we document all our scripts in
`README.md`, so I just added docs for the most salient ones.

* [CHORE] Add CSS build watcher

We watch other stuff on `npm run start`, so we should probably watch CSS
too.