[XUI-885] Update action button spacing (#4476)
This updates action button spacing to match new designs. It reverts a
change made in #4453, which worked well in React-land but caused
unexpected knock-on effects in the HTML/CSS components. I’ve swept up a
couple other small errors highlighted in the KSS docs and updated the
appropriate visual regression reference images.