[XUI-796] Keep tooltip open when tooltip is hovered (#4484)
This partially reverts a change made in #3468 when the `onMouseOver` and
`onMouseOut` methods to open and close the tooltip were moved from the
wrapper to the trigger. Moving these back to the wrapper restores the
desired behaviour. I added a unit test that seems to catch this if we
break it again in the future.

I silenced the linting errors around the focus/blur handlers being on a
separate element from the mouse events because I figured it would be
worse to remove those handlers from an anchor or button that it would to
have them live a small distance apart.