[XUI-905] Unify content block and accordion styles
This follows on from changes made in #4460 to use a content block item
as an accordion trigger. Similar to #4458, I created a mixin for the
styling and moved the KSS docs out to separate files for ease of
bookkeeping. This also required adding a small accordion hook into
`XUIContentBlock` and shuffling around a couple imports in the base CSS
file.

After consulting with design, I also made some styling tweaks, removed
some magic spacing classes from the stories, and changed avatars to the
`small` size to match our new headers. There were also a bunch of
snapshots and visual regression tests to update as well.

Accessibility testing uncovered events in the right content of
`XUIContentBlock` being swallowed by a `preventDefault`. I fixed this
and added a unit test to hopefully catch this in the future. (This
probably fixes [XUI-920](https://jira.teamxero.com/browse/XUI-920).)