[XUI-33] Add ARIA roles to accordion bits for accessibility (#4559)
It’s possible to use the `details` and `summary` elements to achieve
this too. I usually prefer to use the correct semantic element instead
of plastering ARIA roles everywhere; but doing so in this case would
require a polyfill for IE11, which I didn’t feel like bothering with.
These changes come with the requisite snapshot updates.