[XUI-954] Prevent KSS sidebar panels from stacking
We show and hide sample markup panels with an... *interesting* variation
of the radio button CSS hack. We need to manually set the state of all
the relevant radio buttons, closing all the sample markup panels on the
page, before opening the correct one.

I used tabs instead of spaces to keep with the existing format of the
file, but Prettier did trim some whitespace from an earlier comment.