[XUI-999] Sunset ‘link’ button variants
There was no need to add our ‘sunsetting’ flag to the docs, since this
variant didn’t actually appear in them. It did appear in the code
(effectively aliased to `borderless-primary`), so I removed it and
changed its two appearances in banner and toast actions.

Turns out that borderless split buttons were a thing – possible to
create; but as far as I could tell from a GitHub search, nobody was
using them. (Hopefully because it’s a bad idea.) I removed the
possibility of creating them with React and added a codemod just in
case.

I touched up the prop documentation, and our new lists of possible
variants spun up a few new visual-regression tests, the reference images
for which I include here.