[FIX] Fix TS error on build (#4675)
The new version of react-day-picker contained [a regression that failed
to export some needed
types](https://github.com/palantir/blueprint/issues/3790#issuecomment-562321502).
We didn’t notice this because the new package was in `patch` and we
didn’t bring types in until `minor`. This downgrades the package to one
that plays nicely with our setup; or at least that’s what my passing
build locally leads me to believe.