[TEST] Silence TS-related lint warnings
ESLint was giving us a few warnings for explicit `any`s. The ones for
`XUITable` are a result of the fact that the types haven’t been fully
audited yet; and the autocompleter and button ones are legit `any`s,
since we’re not fully TypeScripted yet. Consulted with @dev-alexlapwood,
who assured me ignoring these warnings is okay at this point.