[XUI-1085] Remove redundant margin from XUIBanner (#4725)
`.xui-banner-layout` specified both a width of 100% and auto left and
right margins. If the element is the full width of its container,
there’s not much need to set margins on it to centre it in the
container. (I’m not sure `inline-block` is really what we want for a
banner, but that’s a question for another time.)