[editable] [XUI-1155] Merge tag '16.2.3' into editable (#4751)
* [REVERT] Revert "[XUI-1040] Stub out editable table component"

This reverts commit c5d1571bff12d3712caa681a72fa9aaac62088b4.

We released the editable table stub with 16.2.0, but then questioned
the wisdom of making stub documentation public. This reverts the commit
that brought in the stub, and cherry-picking it back in later will
restore it when we’re ready to go.

* 16.2.1

* Add Amritah's photo to feedback page of the docs (#4689)

* [FIX] Fix npm audit vulnerabilities

On my last `npm install`, I was greeted with this intimidating thing:

```
found 39159 vulnerabilities (39136 low, 23 moderate) in 950168 scanned packages
  run `npm audit fix` to fix 39138 of them.
  7 vulnerabilities require semver-major dependency updates.
  14 vulnerabilities require manual review. See the full report for details.
```

This fixes the low vulnerabilities that came in courtesy of a couple
sub-dependencies.

* XUI-1078 Fixing Tag alignment in Contentblock in FireFox. (#4693)

* [XUI-1091] Add data-automationids to date picker

* [patch] XUI-1075 Adjust truncation behaviour for breadcrumbs and tweak vis-reg (#4688)

* XUI-1075 Update styles for breadcrumbs. Tweak stories

* Updated vis-reg as a result

* Update TS recommendation for copying component prop types (#4686)

* Disable esModuleInterop for TS (#4700)

Using this setting requires our consumers to also use this setting, which is undesirable as it adds a bunch of boilerplate to the build output.

* XUI-1074 Add guard for undefined observer entries (#4701)

* [XUI-971] Improve parsing of KSS example classes

This adds a Handlebars helper for ‘exists in a list’ instead of the
strict-equality helper that was used before. There was a similar
[unused] helper for ‘does not exist in a list’, which I renamed for
clarity and consistency.

* [XUI-971] Avatar KSS docs tweaks

This task was mostly about a markup link, but there were also quite a
few little boo-boos I couldn’t help but clean up while I was in the
file.

* [XUI-971] Prettier changes

There was some weird spacing and mixture of spaces and tabs going on
here, so I converted tabs to spaces and ran Prettier over it to line it
up with the rest of the codebase. `git diff -w` returns empty.

* [TEST] Factor out Chai and Sinon

We were using Chai and Sinon in a few places in our unit tests, so I
pulled them out in favour of Jest and Enzyme, also deleting some tests
that duplicate existing visual regression tests. (I did leave a ReactDOM
test helper in place in a couple button tests due to its better handling
of focus and `document.activeElement`.) I also uncommented a passing
test we were skipping for some reason.

* [TEST] Silence SVG warnings in Jest tests

One of our bar chart subcomponents was rendering SVG child elements
outside of an `<svg>` tag. The tests were passing, but we were getting
some ugly warnings. This silences them.

* [TEST] Silence extraneous Jest-related lint warnings

We were getting some unnecessary Jest-related warnings when running
ESLint, the majority of which were being triggered on non-Jest files. I
disabled the warning in our select box helper.

* [TEST] Silence TS-related lint warnings

ESLint was giving us a few warnings for explicit `any`s. The ones for
`XUITable` are a result of the fact that the types haven’t been fully
audited yet; and the autocompleter and button ones are legit `any`s,
since we’re not fully TypeScripted yet. Consulted with @dev-alexlapwood,
who assured me ignoring these warnings is okay at this point.

* [DOCS] Update link to FT Origami docs style guide

The current link leads to a ‘this link is dead’ page in GitHub – this is
the new and improved version. (I think it’s worth reviewing this section
in general, but that should probably be sized and prioritised with other
work.)

* Version bump for 16.2.2 (#4707)

* [XUI-1067] Upgrade Styleguidist (#4710)

The new version of Styleguidist depends on a new version of
react-docgen. Everything still works as normal locally; let’s see if
this will solve the problem of disappearing props-and-methods buttons in
prod.

* [patch] Update the path for pagination in React doc (#4718)

* [XUI-1126] Update babel/preset-env
Why? This change ensures that the bug with `proposal-numeric-separator` is resolved.
Also, this resolves the issue in [XUI-1067] and the prop buttons now builds correctly on styleguidst.

* [XUI-1126] Package.lock changes for 9100ca29f11e0fbf4425608edceba8bc502c89a5

* [patch] Remove onLabelClick in SelectBox (#4720)

* [XUI-1085] Remove redundant margin from XUIBanner (#4725)

`.xui-banner-layout` specified both a width of 100% and auto left and
right margins. If the element is the full width of its container,
there’s not much need to set margins on it to centre it in the
container. (I’m not sure `inline-block` is really what we want for a
banner, but that’s a question for another time.)

* [patch] XUI-1077 Dropdown in mobile format not unlocking scroll when unloading (#4727)

* Fix Styleguidist for IE 11 (#4729)

Styleguidist > 9.0.4 doesn't work with IE 11

* Bump version to 16.2.3 (#4731)

* 16.2.3 hotfix (#4734)

* [FIX] Remove workaround in Styleguidist sidebar

In #4657, we added a workaround for [Styleguidist breaking our section
links](https://github.com/styleguidist/react-styleguidist/pull/1364).
When we downgraded Styleguidist in #4729, we failed to remove this
workaround, causing the sidebar `href`s to render as `undefined`. This
reverts the offending workaround.