[XUI-1092] Disambiguate links to component writing guide (#4758)
We weren’t super clear about what thing was the product language guide
and what thing was the component writing guide. This changes the link
tooltip on the relevant KSS pages, and I also took the liberty of adding
a link to where we mention it in the ‘displaying text’ section.

The ticket also mentions how we may benefit from making the writing
guide links more obvious, which is a good point; but I’m thinking it’s
out of scope for this task.