[STYLE] Unify heading cell styles
Heading cells were behaving differently based on whether they were in
the header or the footer. Moving the styling to the cell itself makes
them look the same regardless of where in the table they live.

I also moved the calculations we were doing to the background colour and
renamed the variable to match the element. This colour doesn’t appear to
match any of the ones in our basic palette, which is something I’m not
sure if we eventually want to address.