[XUI-888] Add ability to easily add icons to button content
This adds props for adding an icon to the left or right of the text in a
`XUIButton`. This only allows for one or the other, protected by a
console warning and a custom prop validator, except in the case of a
caret. Speaking of carets, since what is a caret if not an icon to the
right of the text in a `XUIButton`, I removed the separate
`XUIButtonCaret` component in favour of the prop `hasCaret` on
`XUIButton`. Lots of small vis-regs and test adjustments with this one.

In enabling icons for `small` and `xsmall` buttons, I created `small`
and `xsmall` icon size classes. For legibility, these only change the
size of the wrapper – the icon itself will never get smaller than it is
in a `medium` wrapper. I also de-magic-numbered the button line-height –
looks like we already had the var, we just weren’t using it anywhere.

IE11 wasn’t playing well with Flexbox when launching a dropdown from a
secondary button, so I added a class as a styling hook. This affected a
bunch of snapshots, but the vis-reg tests didn’t note any changes.