[REFACTOR] Update built-in Sass functions to use modules
Sass’s move to module format includes changes to how we call [existing
built-in
functions](https://github.com/sass/sass/blob/master/accepted/module-system.md#built-in-modules-1).
This updates the existing functions in our codebase to use the new
module syntax. There are a few deprecated functions, of which the only
one we use is `transparentize`, which has been updated here.

It is possible to add a module to the root namespace on a per-file basis
– in English, that means saying `@use 'sass:color' as *` allows us to
use `mix($foo, $bar)` instead of `color.mix($foo, $bar)`. This has the
potential for abuse and confusion, so I’m only using it in places like
`_colors.scss` where we (a) only use one module that’s (b) the fairly
obvious one for the file. Elsewhere, I opted for the explicit syntax for
clarity and safety.