[XUI-1064] Add ‘XUI’ prefix to dropdown components
This is basically a one-for-one replacement, converting `DropDown` and
friends to `XUIDropDown` and friends. We are maddeningly inconsistent in
how we refer to components in the docs, so I tidied up where I could.

Some vis-reg-related stuff from #4884, including the reference image,
was done against an older version of `breaking-changes`. Since it’s
dropdown-related, I’ve treated it like a merge conflict and resolved it
here.