[XUI-1383] Fix casing in dropdown-related components
Our React components were called variations of Pascal-cased `DropDown`,
which put us at odds with general design-system practice, as well as our
very own [XUI
Guide](https://xui.xero.com/16.5.0/section-components-collectinginput-dropdowns.html).
This extends the work in #4893 to change occurrences of `DropDown` to
`Dropdown`, not only in the component names, but in the associated
public methods as well.

This also restores a couple sub-sections to Styleguidist that I don’t
think made it through the previous round of changes.