[XUI-1384] Update files affected by Prettier changes
We updated Prettier and the associated config in #4873. This appears to
have introduced a couple small changes to the rules that Prettier began
to flag. This is the result of the automated clean-up: lots of
line-length changes, trailing semicolons in inline styles, and spaces
after the `function` keyword, mostly.

This change avoids introducing these changes one-by-one when we touch
these files, keeping our diffs cleaner.