[XUI-1064] Add ‘XUI’ prefix to select box and friends
If you haven’t seen it before, it’s new to you! Added the prefix and
removed the hyphen from the directory name, as well as made a few
ancillary doc-style-related tweaks.

The directory rename let me remove a piece of logic in the Styleguidist
config, but another piece is still broken – I’ve opened
[XUI-1411](https://jira.teamxero.com/browse/XUI-1411) to fix it.