[XUI-1419] Fix spacing in inline reversed input groups
Label margins were broken in inline reversed checkbox and radio groups.
Since they aren’t technically `XUICheckboxGroup`s or `XUIRadioGroup`s,
we didn’t have a lot of non-ugly ways to target them. We’ve opened
[XUI-1426](https://jira.teamxero.com/browse/XUI-1426) to rethink how we
handle these, which will include another look at styling. This is merely
a stopgap to fix the immediate breakage. Since I only caught this with
my eyeballs, I’ve added visual regression tests to catch this situation
in the future.

Speaking of which, this also includes an update to a vis-reg reference
image from one of our kitchen-sink tests. It doesn’t represent a
realistic use of the component, so a change in the vis-reg was
deemed acceptable.