[XUI-1471] Update Artifactory URLs
We were seeing unexpected `package-lock.json` changes when running an
`npm install`. Turns out that our `.npmrc` was pointing to a URL that
redirected to the canonical Artifactory URL. At the advice of the
Artifactory team, we’re changing `.npmrc` to point to the OG URL instead
of the redirect.

Unblocks #4967.