[XUI-828] Add ESLint and Prettier scripts and clear out .ignore files (#5023)
ESLint and Prettier currently have no way to specify multiple .ignore files which means we have to duplicate .gitignore in .eslintignore and .prettierignore.

For ESLint we can get around this by using the --ignore-pattern flag. Rather than adding the flag to every place we call ESLint, I've added a new script to do it for us.

For Prettier we can do the same thing, but instead of using --ignore-pattern we add paths starting with '!'. e.g. '!scripts'.

These scripts do not make use of our task runner because ESLint and Prettier appear to do some funky things with stdout that are not compatible.