[XUI-1554] Fix banner overflow
I broke this by setting `display: block` in #4999. Changing the
`display` messed with the overflow, and fixing the overflow messed with
the sentiment borders. I left the border compensation gubbins in the
mixin in case we need it in the future – happy to blow it out if y’all
feel it’s necessary.

I also corrected a var spelling that was bothering me.