Skip Prettier if running on a specific file that should be ignored (#5127)
This fixes an issue where prettier would error out during lint-staged. This happened because lint-staged calls Prettier with each staged file as an argument, and Prettier errors out if it is called with a file for an argument in tandem with an exclude glob.

eg. prettier versions.js !versions.js