[XUI-1651] Remove deprecated Storybook addons
This removes the unused addon `storybook-readme` (it had been commented
out pending an IE11 fix, and IMHO there’s no point doing work to
reenable it if we’re redoing the docs soon anyway) and replaces the
deprecated `addon-centered` in favour of `addParameters: 'centered'`
when possible, or our old `xuiResponsiveCenter` decorator if not. A
couple formatting tidy-ups made their way in here too.