[XUI-1651] Update and modify Storybook decorators
This updates our decorators to be compatible with Storybook v6 and also
adds a new default decorator to provide styling similar to the existing
sandbox, mainly to prevent having to manually review a ton of new
vis-reg images. I also tweaked the existing responsive decorator (still
in use in several places). This new default decorator allowed me to swap
out our custom decorator for default Storybook behaviour or remove the
decorator completely.

It also allowed me to back out a lot of the explicit `selectors` I added
to the Backstop config. [It’s not clear whether `addDecorator` at a test
level is on its way to being phased
out](https://github.com/storybookjs/storybook/blob/next/MIGRATION.md#deprecated-addparameters-and-adddecorator),
but this is the best I could do without having to make this diff even
larger than it already is.