[XUI-1651] Text input story changes
Tweaking the name of the test caused name change in the reference images
– the price we pay for tidiness. Also, Storybook will occasionally
render the components before the main sandbox is finished rendering,
breaking `box-sixing` inheritance. I included it explicitly to un-break
these tests. I also modified the tests to conditionally add a decorator
to the responsive stories to keep them passing with the existing
reference images.