[XUI-1651] Add resize observer to autocompleter to fix tests
To render properly, autocompleter with non-wrapping pills depends on the
width of the placeholder. This width was initially getting calculated
before the CSS was applied, causing an incorrect measurement and the
placeholder to be cut off in some cases. Is this likely to happen in the
Real World™?  Not sure! This adds a resize observer to make sure the
width is calculated correctly regardless.