[XUI-1809] Automate datepicker localisation
This removes hard-coded strings and reliance on potentially incorrect
passed-in props in favour of localising month and day names, text
direction, and first day of the week based on locale. It follows that
`locale` is now required, and `nextButtonAriaLabel` and
`nextButtonAriaLabel` are also required now to align with the way we
handle labels across the rest of XUI. This also includes documentation
and enough changes to tests so that the existing tests pass.