[patch] Use the Node API for Webpack to fix missing build output (#5310)
Using the CLI from within Node limits our ability to debug issue and control the console output.

I suspect this was originally done before there was a Node API for Webpack, but now that there is one we should use it.

The main advantage for us is that by controlling the output of the build stats we can prevent webpack's Ascii control characters from breaking our console output. This means that running `npm run build` will now output the build log in the console correctly.