[XUI-1809] Add tests for new auto-localisation functionality
I used React Testing Library for the new unit tests and also ported over
the few that were already there. This also installs and configures
[`full-icu`](https://github.com/unicode-org/full-icu-npm), which is
required for Node to have access to the localised strings it needs.
(Could we take out the unit tests here and handle the localisation
exclusively with vis-regs?)