[patch] [XUI-2030] Ensure XUIDateRangeInput onSelectDate methods are called (#5450)
* [XUI-2030] Make DateRangeInput examples controlled components

As most of our consumers will likely want to do something with the date provided, it makes sense for our React docs examples to be controlled components. This helps for copy/paste implementations, and these examples now function within the docs.

* [XUI-2030] Ensure onSelectDate props are not passed directly into XUIDateInputItems

As props were passed in with the onSelectDate name, they were overriding the manual definition of the prop within XUIDateRangeInput which referred to the local method.
This commit ensures that they no longer override the desired method.