[XUI-2008] Add component generator
This generates a file skeleton for a new XUI component and also modifies
the appropriate existing config files. It’s not _quite_ smart enough to
put everything in its right place, so I detailed the few necessary
manual steps in `CONTRIBUTING.md`. I elected to run
[plop](https://plopjs.com/), the utility that does the heavy lifting
here, from npx mostly to keep it out of the permanent dependency graph
since this will ideally not be used super-often.