[FIX] Remove unused security checks
This sounds a little scary, but the package the script calls was removed
three years ago (#3678). The ‘What are we going to do about this?’
comment dates from the XUI 14 alpha slightly thereafter (#3742), which
leads me to believe that we, in fact, did not want to do anything about
it. (This doesn’t rule out doing anything about it in the future,
though!)