[patch] [XUI-2600] Manage cell text selection via onClick (#5724)
* [XUI-2600] Add support for focusing end of multiline input

This removes our manual list of supported types, and instead checks if the method exists on the input.

* [XUI-2600] Add interactive creator row example to storybook

As we support products that commonly use the creator row pattern, it may be helpful to have a performant example within storybook of how we recommend this pattern be implemented. This example is copied from the React docs example, but should be more performant for testing, and is a more reliable reproduction of behaviour.

* [XUI-2600] Manage select via focus and mousedown combination

As the native behaviour of tabbing between cells is to select the contents, we can instead use onClick and onMouseDown to ensure selection works as expected.

* [XUI-2600] Update test names

* [XUI-2600] Rename `isClicking` to `isFocusingViaClick`

* [XUI-2600] Add tests for selection edge cases

(cherry picked from commit 9f626f0a76f9c04dcab4456bbc5930c1568b0e4e)