[noref] Adjust PR Trigger (#5846)
This should make our trigger more like the old implementation, which will result in fewer builds. The default implementation resulted in a rebuild occurring each time the base branch *or* the compare branch were changed, technically providing the most up-to-date build but resulting in a lot of rebuilding.
This change should mean that rebuilds will only be triggered when github recalculates the commit hash, which is supposedly less often but should be our old implementation.