[XUI-2531] Disable `aws.region` requirements (#5924)
The recommended approach for pipelines is to have the agent used for deploying to operate in the same region as that which it is deploying to. Generally, this is to avoid cross-region fees and increase speed.
Unfortunately, as we are operating in `teamcity1`, we do not have access to `us-east-1` agents.
Forunately, this is considered less of an issue for frontend projects, as the artifacts are passed between builds.

This commit removes the default requirement for the agent region to match the deployment region, meaning we should be able to use our existing agents.