[XUI-2531] Update PR trigger branchFilter (#5947)
Unfortunately, it appears that the `branchFilter` is set based on logical name (what appears in the TeamCity UI) rather than branch specification, meaning that it is not easy to filter exclusively to PR branches, as these appear as just the PR number itself. Instead, we need to filter out our explicit non-PR branches.