[patch] [XUI-3079] Fix type defs for components with spreadProps (#5957)
* Fix issue with `Omit` type over union types
* Re-pattern all SpreadProps types and add comment to help prevent
  making the same mistake again via copy-pasta
* (unrelated) Add an eslint exemption to silence a linting warning

A previous type fix introduced a type bug. Previous commit message:

> When spread props are merged with a component’s base props, the keys
> of the base props should be omitted from the spread props to
> accurately reflect the JS usage.

9b83338f17167ed49b5fd757fcaed2bfd0c92c2b

When a union type is used as the base for an `Omit` type, the union type
will only include the union of the types in the union (a set operation).
This means that types that do not appear in all types of the union will
be omitted by the `Omit` type which is not the desired behaviour. This
surfaced as an issue where a `XUIButton` component would no longer
accept a `form` prop.

The solution is to add the `Omit` to every type within the union, so
that the union type can behave as it did before.