[XUI-2765] Fix undefined ariaRole for XUIPickitems (#5959)
Previously, `XUIPickitem` elements within a `XUIPicklist` would
have their `ariaRole` set by the `XUIPicklist`. If the `XUIPicklist` had
a role of `listbox`, the `XUIPickitem` would be given a role of
`option`. Otherwise, it would be set to `undefined`. This occurred
regardless of whether a custom `ariaRole` had been provided to the
`XUIPickitem`.

This change means that the `option` role will still be given to
`XUIPickitem`s within a `listbox` role `XUIPicklist`, but will otherwise
fall back to the `XUIPickitem`'s `ariaRole`.