[XUI-3115] Update date input and date range input stories
I removed the fancy conditional gubbins that toggles between the two in
favour of treating them as two separate components. This required some
surgery to the variations file, as well a tweak to the vis-reg testing
config and renaming the reference images.

I also made some changes to the suggested date helpers based on design
feedback and did a little refactoring to make some things easier to
understand. This did involve a little more use of Day.js, but I think it
was worth it.